refactor: 테스트 코드들의 패키지 정리
