+++ src/test/java/racingcar/domain/car/CarTest.java
@@ -1,8 +1,6 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.Test;
-import racingcar.domain.car.Car;
-import racingcar.domain.car.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
